fix: TextArea warning with React 16.13 (#21703)
* fix: TextArea auto height warning

* comment it

* popout textarea test case

* no warning in react 16.13
